Fix edgecase reopening bug.
When users re-open lots of closed tabs rapidly (e.g. quit chrome while many journals are open,
restart, re-open chrome), sometimes the tabs beat the XHR to check the network status (or if
the network has a slow DHCP server, it might take ~30 sec or more).

This pulls the last known network state from LSOs and uses that as a guess before the XHR.